+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Custom labels | Remove a dashes below labels. (UIEH-1361)
 * Titles tab: Add a Packages facet. (UIEH-1350)
 * Title detail record: Packages accordion - Honor Packages Facet selection(s). (UIEH-1351)
+* Fix errors appearing when user switch to "User consolidation" pane. (UIEH-1362)
 
 ## [8.0.3] (https://github.com/folio-org/ui-eholdings/tree/v8.0.3) (2023-03-30)